python project/requirements.txt
opencv-python==4.8.1.78
ultralytics==8.0.196
pyttsx3>=2.90
Pillow>=9.0.0
numpy>=1.20.0
